requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
docker==7.0.0
idna==3.6
iniconfig==2.0.0
mirakuru==2.5.2
packaging==24.0
pluggy==1.4.0
port-for==0.7.2
psutil==5.9.8
pymongo==4.6.3
pytest==8.1.1
requests==2.31.0
testcontainers-core==0.0.1rc1
testcontainers-mongodb==0.0.1rc1
urllib3==2.2.1
wrapt==1.16.0
